Refactoring: Move all the logic to retrieve mixes information from background to popup
