+++ src/background.js
@@ -26,8 +26,6 @@ chrome.webRequest.onBeforeRequest.addListener(graphQLListener,
   { urls: ['https://app.mixcloud.com/graphql'] }, ['requestBody']
 )
 
-chrome.runtime.onMessage.addListener(onMessageListener)
-
 async function graphQLListener (spiedRequest) {
   if (spiedRequest.requestBody) {
     const byteArray = new Uint8Array(spiedRequest.requestBody.raw[0].bytes)
@@ -129,86 +127,4 @@ async function storeCloudcast (cloudcast, usernameAndSlug) {
     console.log('savecloudCast ' + dataToStore.path)
     store.setMixData(dataToStore)
   }
-}
-
-function onMessageListener (request, send, sendResponse) {
-  if (request.action === 'getTracklist') {
-    chrome.tabs.query({ url: '*://*.mixcloud.com/*' }, async (tabs) => {
-      if (tabs.length > 0) {
-        const pathsAndTitle = tabs.map((tab) => {
-          const url = tab.url
-          const title = tab.title
-          const regex = /^\D*:\/\/\D+\.mixcloud\.com/
-          return {
-            url: decodeURIComponent(url.replace(regex, '')),
-            title: title
-          }
-        })
-        const mixesDataFromStore = new Promise((resolve, reject) => {
-          return getMixesData(pathsAndTitle.paths, 1, resolve, reject)
-        })
-
-        let isMixPathFromPlayerFound = false
-        let mixPathFromPlayer
-        for (let i = 0; i < tabs.length && !isMixPathFromPlayerFound; i++) {
-          mixPathFromPlayer = await getMixPathFromPlayer(tabs[i])
-          isMixPathFromPlayerFound = !!mixPathFromPlayer
-        }
-
-        try {
-          const mixesData = await mixesDataFromStore
-          const mixesDataforPopUp = {
-            currentPlayedMixPath: mixPathFromPlayer,
-            mixesData: mixesData
-          }
-          console.log('data sent to popup')
-          console.log(mixesDataforPopUp)
-          sendResponse(mixesDataforPopUp)
-        } catch (e) {
-          console.error('Error on getTracklist : ' + e)
-          sendResponse()
-        }
-      } else {
-        sendResponse()
-      }
-    })
-    return true
-  }
-}
-
-/**
- * Recursive function to handle asynchronicity between request's spy & tracklists display
- * @param path : mix path
- * @param counter : attempt counter for tracklist retrieval from store
- * @param resolve
- * @param reject
- * @returns {*} resolve(an array of mixesData or empty array)
- */
-function getMixesData (paths, counter, resolve, reject) {
-  if (counter > 3) {
-    resolve([])
-  }
-  store.getMultipleMixData(paths).then((mixesData) => {
-    if (mixesData.length === 0) {
-      setTimeout(function () {
-        getMixesData(paths, counter + 1, resolve, reject)
-      }, 500)
-    } else {
-      store.getMultipleMixData(paths).then((mixesData) => resolve(mixesData))
-    }
-  })
-}
-
-async function getMixPathFromPlayer (tab) {
-  return new Promise((resolve) => {
-    chrome.tabs.sendMessage(
-      tab.id,
-      {
-        action: 'requestMixPathFromPlayer'
-      },
-      (response) => {
-        resolve(response)
-      }
-    )
-  })
 }
